+++ game/plugin/skills/mining/src/gem.kt
