+++ game/plugin/skills/mining/src/mining.plugin.kts
@@ -1,0 +1,189 @@
+import org.apollo.cache.def.ItemDefinition
+import org.apollo.game.action.AsyncDistancedAction
+import org.apollo.game.action.DistancedAction
+import org.apollo.game.message.impl.ObjectActionMessage
+import org.apollo.game.model.Position
+import org.apollo.game.model.entity.Entity
+import org.apollo.game.model.entity.EquipmentConstants
+import org.apollo.game.model.entity.Player
+import org.apollo.game.model.entity.Skill
+import org.apollo.game.model.entity.obj.StaticGameObject
+import org.apollo.game.plugin.skills.mining.*
+import org.apollo.game.scheduling.ScheduledTask
+import java.util.*
+
+class MiningAction(val player: Player, val objectID: Int, val p: Position, val ore: Ore) : AsyncDistancedAction<Player>(PULSES, true, player, p, ORE_SIZE) {
+
+    companion object {
+        private val PULSES = 0
+        private val ORE_SIZE = 1;
+    }
+
+    private val rand: Random = Random()
+
+    override suspend fun executeActionAsync() {
+        val level = mob.skillSet.getSkill(Skill.MINING).currentLevel
+        val pickaxe = findPickaxe()
+
+
+        //check that our pick can mine the ore
+        if (pickaxe == null || level < pickaxe.level) {
+            mob.sendMessage("You do not have a pickaxe for which you have the level to use.")
+            stop()
+            return
+        }
+
+        //check that we can mine the ore
+        if (level < ore.level) {
+            mob.sendMessage("You do not have the required level to mine this rock.")
+            stop()
+            return
+        }
+
+        //start the process of mining
+        wait(pickaxe.pulses, {
+            mob.sendMessage("You swing your pick at the rock.")
+            mob.playAnimation(pickaxe.animation)
+        })
+
+        if (miningSuccessful()) {
+
+            //Check inv capacity
+            if (mob.inventory.freeSlots() == 0) {
+                mob.inventory.forceCapacityExceeded()
+                stop()
+                return
+            }
+            if (mob.inventory.add(ore.id)) {
+                //TODO: Use lookup from utils once it has a lookup function for IDs
+                val oreName = ItemDefinition.lookup(ore.id).name.toLowerCase();
+                mob.sendMessage("You managed to mine some $oreName.")
+                mob.skillSet.addExperience(Skill.MINING, ore.exp)
+                //Expire ore
+                var rockEntity: StaticGameObject? = null
+                val region = mob.world.regionRepository.fromPosition(position)
+                val entities = region.getEntities(position)
+                for (entity: Entity in entities) {
+                    if (entity is StaticGameObject) {
+                        if (entity.id == objectID) {
+                            rockEntity = entity
+                        }
+                    }
+                }
+                if (rockEntity == null) { //Mining entity not found at location...
+                    stop()
+                    return
+                }
+
+
+                //Get ID of exipred ore
+                val expiredObjectID = ore.objects.get(objectID);
+                val expiredRockEntity = StaticGameObject(mob.world, expiredObjectID!!, position, rockEntity.type, rockEntity.orientation)
+                //add task to remove normal ore and replace with depleted
+                mob.world.schedule(object : ScheduledTask(0, true) {
+                    override fun execute() {
+                        //Replace normal ore with expired ore
+                        region.removeEntity(rockEntity);
+                        region.addEntity(expiredRockEntity)
+                        this.stop() //Makes task run once
+                    }
+                })
+                //add task to respawn normal ore
+                mob.world.schedule(object : ScheduledTask(ore.respawn, false) {
+                    override fun execute() {
+                        //Replace expired ore with normal ore
+                        region.removeEntity(expiredRockEntity)
+                        region.addEntity(rockEntity);
+                        this.stop() //Makes task run once
+                    }
+                })
+            } //If we did not add to inv, the action will still stop
+            stop(); //Force this action to stop after we are done
+        }
+
+    }
+
+    private fun findPickaxe(): Pickaxe? {
+        for (pick in getPickaxes()) {
+            if (pick.level > mob.skillSet.getSkill(Skill.MINING).currentLevel) {
+                continue;
+            }
+            val weaponSlot = mob.equipment.get(EquipmentConstants.WEAPON)
+            if (weaponSlot != null && weaponSlot.id == pick.id) {
+                return pick;
+            } else if (mob.inventory.contains(pick.id)) {
+                return pick;
+            }
+        }
+        return null;
+    }
+
+
+    /**
+     * Returns the chance of mining being successful.
+     * Algorithm comes from: http://runescape.wikia.com/wiki/Talk:Mining#Mining_success_rate_formula
+     */
+    private fun miningSuccessful(): Boolean {
+        val level = mob.skillSet.getSkill(Skill.MINING).currentLevel
+        val percent: Double
+        if (ore.chanceOffset) {
+            percent = (ore.chance * level + 1) * 100
+        } else {
+            percent = (ore.chance * level) * 100
+        }
+        return rand.nextInt(100) < percent;
+    }
+}
+
+class ExpiredProspectingAction : DistancedAction<Player> {
+
+    constructor(mob: Player, position: Position) : super(PROSPECT_PULSES, true, mob, position, ORE_SIZE)
+
+    companion object {
+        private val PROSPECT_PULSES = 0
+        private val ORE_SIZE = 1;
+    }
+
+    override fun executeAction() {
+        mob.sendMessage("There is currently no ore available in this rock.")
+        stop();
+    }
+}
+
+class ProspectingAction(val m: Player, val p: Position, val ore: Ore) : DistancedAction<Player>(PROSPECT_PULSES, true, m, p, ORE_SIZE) {
+
+    companion object {
+        private val PROSPECT_PULSES = 3
+        private val ORE_SIZE = 1;
+    }
+
+    var started = false;
+
+    override fun executeAction() {
+        if (started) {
+            val oreName = ItemDefinition.lookup(ore.id).name.toLowerCase()
+            mob.sendMessage("This rock contains " + oreName + ".")
+            stop();
+        } else {
+            started = true
+            mob.sendMessage("You examine the rock for ores...")
+            mob.turnTo(position)
+        }
+    }
+}
+
+on { ObjectActionMessage::class }
+        .where { option == 1 }
+        .then {
+            if (lookupOreRock(id) != null) {
+                it.startAction(MiningAction(it, id, this.position, lookupOreRock(id)!!))
+            }
+        }
+
+on { ObjectActionMessage::class }
+        .where { option == 2 }
+        .then {
+            if (lookupOreRock(id) != null) {
+                it.startAction(ProspectingAction(it, this.position, lookupOreRock(id)!!))
+            }
+        }
