+++ game/plugin/skills/mining/src/ore.kt
